% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2hap.R
\name{get_hap}
\alias{get_hap}
\title{generat haps from vcf}
\usage{
get_hap(vcf, hap_prefix = "H",
filter_Chr = F, Chr = "scaffold_1",
filter_POS = F, startPOS = as.numeric(), endPOS = as.numeric(),
hyb_remove = T, na.drop = T)
}
\arguments{
\item{vcf}{vcf}

\item{hap_prefix}{hap_prefix}

\item{filter_Chr}{filter_Chr}

\item{filter_POS}{filter_POS}

\item{hyb_remove}{hyb_remove}

\item{na.drop}{na.drop}
}
\description{
generate hap format from vcf

import vcf file
}
\author{
Zhangrenl
}
