% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2hap.R
\name{hap_result}
\alias{hap_result}
\title{generate hap results}
\usage{
hap_result(hap, out = T, file = "hapResult.txt")
}
\arguments{
\item{hap}{hap}

\item{out}{write hap results to a txt file}

\item{file}{file path}
}
\description{
summarize hap result and output a txt file
}
