% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenos.R
\name{hapVsPheno}
\alias{hapVsPheno}
\title{hapVsPheno}
\usage{
hapVsPheno(hap,
pheno,
phenoName,hapPrefix = "H",
geneID = "Seita.1G000000",
mergeFigs = TRUE,
minAcc = 5)
}
\arguments{
\item{hap}{hap}

\item{pheno}{pheno}

\item{phenoName}{pheno name}

\item{hapPrefix}{hap  prefix}

\item{geneID}{gene ID for plot title}

\item{mergeFigs}{merge heatmap and box plot or not}

\item{minAcc}{If observations numberof a Hap less than this number will
not be compared with others or ploted}
}
\description{
hapVsPheno
}
