% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenos.R
\name{hapVsPhenos}
\alias{hapVsPhenos}
\title{hapVsPhenoS}
\usage{
hapVsPhenos(hap, phenos,hapPrefix = "H",geneID = "Seita.0G000000", mergeFigs = T)
}
\arguments{
\item{hap}{hap}

\item{phenos}{phenos}

\item{hapPrefix}{hap  prefix}

\item{geneID}{gene ID forplot title}

\item{mergeFigs}{merge heatmapand box plot ornot}
}
\description{
hapVsPhenoS
}
