% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_pheno}
\alias{import_pheno}
\title{imports phenos from file}
\usage{
import_pheno(phenoFile, comment.char = "#", ...)
}
\arguments{
\item{phenoFile}{pheno file path, should be a table separated by tab}

\item{comment.char}{comment.char, start with comment.char will be ignored}

\item{...}{parameters will pass to read.delim}
}
\description{
第一列为Accession
第二列起为表型，
第一行为表头（表型名称）：“表型名称（计量单位）.地点[年份]”
}
