% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_vcf}
\alias{import_vcf}
\title{import vcf from file}
\usage{
import_vcf(vcf_file = vcf_file, ...)

import_vcf(vcf_file = vcf_file, ...)
}
\arguments{
\item{vcf_file}{file path of vcf}

\item{...}{pass to vcfR::read.vcfR}
}
\description{
read and import vcf file

import vcf file
}
\author{
Zhangrenl
}
