% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapVisualization.R
\name{plotGeneStructure}
\alias{plotGeneStructure}
\title{plotGeneStructure}
\usage{
plotGeneStructure(gff, hapResult, Chr, startPOS, endPOS)
}
\arguments{
\item{gff}{gff}

\item{hapResult}{ouput of hap_result(), a data.frame consistant
with 4 metarows: Chr, POS, Allele and ANN, and each genotype of each hap}

\item{Chr}{Chr, it will use the first Chr in the meta rows by defalt}

\item{startPOS}{startPOS, it will use the min position in the second meta rows by defalt}

\item{endPOS}{endPOS, it will use the max position in the second meta rows by defalt}
}
\description{
plotGeneStructure
}
