% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2hap.R
\name{get_hap}
\alias{get_hap}
\title{generat haps from vcf}
\usage{
get_hap(vcf, hap_prefix = "H",
filter_Chr = F, Chr = "scaffold_1",
filter_POS = F, startPOS = as.numeric(), endPOS = as.numeric(),
hyb_remove = T, na.drop = T)
}
\arguments{
\item{vcf}{vcf}

\item{hap_prefix}{hap_prefix, defalt is "H"}

\item{filter_Chr}{filter vcf by Chrome or not, defalt is FALSE. If TRUE, the Chr is needed.}

\item{Chr}{Needed for filter vcf by Chrom}

\item{filter_POS}{filter vcf by Position or not, defalt is FALSE. If TRUE, startPOS and endPOS are needed.}

\item{startPOS, endPOS}{Needed for filter vcf by position. startPOS must less than endPOS}

\item{hyb_remove}{Remove accessions contains hybrid site or not. Defalt is TRUE.}

\item{na.drop}{Drop Accessions contains unknown allele site or note. Defalt is TRUE}
}
\description{
generate hap format from vcf

import vcf file
}
\author{
Zhangrenl
}
