% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenos.R
\name{hapVsPheno}
\alias{hapVsPheno}
\title{hapVsPheno}
\usage{
hapVsPheno(hap, pheno,phenoName,hapPrefix = "H", geneID = "Seita.1G000000", mergeFigs = T, minPhenoNumber = 5)
}
\arguments{
\item{hap}{hap}

\item{pheno}{pheno}

\item{phenoName}{pheno name}

\item{hapPrefix}{hap  prefix}

\item{geneID}{gene ID forplot title}

\item{mergeFigs}{merge heatmapand box plot ornot}
}
\description{
hapVsPheno
}
